powerpc: book3e: Protect PID from being "stolen" if the PID is in use by some other device.
Original patch by:
  Michael Ellerman <michael@ellerman.id.au>
  Benjamin Herrenschmidt <benh@kernel.crashing.org>

Currently we will steal a context from a task when we run out of
availble PID numbers in the PID register of the processor.  Some SoCs
have designed devices, usually IOMMUs, that are able to be "given" PID
values so they can run in the context of the task.  While these
devices are grated access to that context we have the requirement to
protect that context from being stolen.

This patch provides interfaces that allow this protection to occur.

Signed-off-by: Jimi Xenidis <jimix@pobox.com>