powerpc: book3e: Add Chroma as an new WSP/PowerEN platform.
This patch add the Chroma platform to WSP/PowerEN, which is a PCIe
card.  The card includes an H8 service processor that can is used to
manage the card.  The H8 is connected over the second serial UART on
the PowerEN chip so this patch includes a simple 16550 driver to
enable communication, mostly for "power off" and "rebooting".

This patch also includes a "halt" method that will shut of all A2
cores but still leave power on.  This is desirable especially if you
wish to interrogate the chip with a hardware probe after the halt.

Signed-off-by: Jimi Xenidis <jimix@pobox.com>