powerpc: book3e: Add arch_exit_map() for mmu_context_nohash to support coprocessors.
This allows software to inc/dec mm_count at the right time.  Using
destroy_context() is insufficient because that is only called when the
reference has gone to zero. However, arch_exit_mmap() is called once
all users of the mm have gone, but before the mm is destroyed - this
is the right point to drop our reference and mark the mm context as
dead.

Signed-off-by: Jimi Xenidis <jimix@pobox.com>