poweren_ep: Host PowerEN PCIe EP base device driver
This is the implementation of a PCI Express driver for
the PowerEN Processor Endpoint. It is intended to allow
drivers that needs PCIe resources to register against it
and ask for them. It also provides a component named
"slot manager" that allocates and configures two 1MB buffers
per function for inbound and outbound mmio transfers.
The manager split the buffers in slots for multi-process and
multi-thread usage. It also provides an iommu memory space manager

Other contributors:
            Michael Barry <mgbarry@linux.vnet.ibm.com>
            Owen Callanan <owencall@linux.vnet.ibm.com>
            Antonino Castelfranco <antonino@linux.vnet.ibm.com>
            Bernard Gorman <berngorm@ie.ibm.com>
            Jack Miller <jack@codezen.org>
            Jimi Xenidis <jimix@pobox.com>

Signed-off-by: Michael Barry <mgbarry@linux.vnet.ibm.com>
Signed-off-by: Owen Callanan <owencall@linux.vnet.ibm.com>
Signed-off-by: Antonino Castelfranco <antonino@linux.vnet.ibm.com>