poweren_ep: PowerEN PCIe EP High Performance Device Driver
This is the implementation of a char device that enables
host and device to exchange data using MMIO and DMA
transfers. It also has a memory registration component
that allow to reserve DMA windows for multi-process and
multi-thread usage. On the host it also allows user
applications to map memory buffers allocated in userspace
against the host iommu.

Other contributors:
                Michael Barry <mgbarry@linux.vnet.ibm.com>
                Owen Callanan <owencall@linux.vnet.ibm.com>
                Antonino Castelfranco <antonino@linux.vnet.ibm.com>
                Bernard Gorman <berngorm@ie.ibm.com>
                Jack Miller <jack@codezen.org>
                Jimi Xenidis <jimix@pobox.com>

Signed-off-by: Michael Barry <mgbarry@linux.vnet.ibm.com>
Signed-off-by: Owen Callanan <owencall@linux.vnet.ibm.com>
Signed-off-by: Antonino Castelfranco <antonino@linux.vnet.ibm.com>