poweren: Virtual Ethernet Driver
PowerEN is an accelerator card that attaches to its host machine via the
PCIe bus. A full operating system runs on the card; the host can offload
work to PowerEN, or the card may preprocess network data before delivery
to the host. This driver exposes a virtio-based virtual ethernet
interface, used to allow communication between the host and PowerEN via
an underlying PCIe driver. Packets are transferred from host to device
over MMIO, and from device to host over DMA. Input SKBs allocated by
virtio_net on the host are "pinned" using the kernel's iommu_map API,
and then "posted" to a buffer queue on the device. The buffers are
subsequently popped off the queue and used as the DMA target address
when the device wishes to send a packet to the host.

    Additional contributors:
      Mark Purcell <mark_purcell@ie.ibm.com>
      Michael Barry <michael_barry@ie.ibm.com>
      Antonino Castelfranco <antonino_castelfranco@ie.ibm.com>

This patch adds all the headers required by the driver. Aside from
function declarations, of particular note are the following:

- poweren_cop.h: co-processor struct definitions and macros, used to
construct commands for submission to the DMA engine and check the
validity of the codes it returns.

- poweren_veth_irq.h: provides interrupt management functions. On the
host, this consists of an RX handler function invoked when the device
raises a host-side interrupt after packet transmission. On the device, a
soft-IRQ polling mechanism is implemented to detect and process incoming
packets.

- poweren_veth_iommu.h: provides a function which identity-maps the
physical page in which a given socket buffers resides to the IOMMU,
should no such mapping currently exist. This IOMMU address will
subsequently be posted to the device, as described above.

- poweren_virtio.h: defines the struct of high-level handler functions
which allow each component of the driver to operate transparently; this
will be populated with different functions for the host and device, e.g.
xmit_pkt will write a full packet to the device's MMIO queue on the
host, while the same function will pop a buffer and submit a DMA request
on the device.

- poweren_vring.h: low-level functions for interacting with a
virtio_ring; adding buffers, retrieving buffers, checking availability,
indicating interruptibility, etc.

Signed-off-by: Bernard Gorman <berngorm@ie.ibm.com>